core/requirements.txt
alabaster==1.0.0
annotated-doc==0.0.4
anyio==4.12.1
asgiref==3.11.1
babel==2.18.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docutils==0.22.4
feedparser==6.0.12
filelock==3.25.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
imagesize==1.4.1
indic-nlp-library==0.92
inflection==0.5.1
Jinja2==3.1.6
joblib==1.5.3
lxml==5.4.0
lxml_html_clean==0.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
Morfessor==2.0.6
mpmath==1.3.0
networkx==3.6.1
newspaper4k==0.9.5
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pgvector==0.4.2
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-file==3.0.1
rich==14.3.3
roman-numerals==4.1.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
setuptools==70.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.8.3
Sphinx==9.1.0
sphinx-argparse==0.5.2
sphinx_rtd_theme==3.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tldextract==5.3.1
tokenizers==0.22.2
torch==2.10.0+cpu
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
w3lib==2.4.0
